Add tooling required for building hydrogen themes
These are the same files as in hydrogen, but with the extension changed to .js instead of .mjs.
Our environment doesn't seem to allow importing .mjs files as they're not interpreted as CommonJS modules.

Some minor changes were made to the files to e.g. adapt them to our directory structure.